added patch ability, code formatting and restructure
